refactor(inference): carve the serving guard and the lane send out of generate_stream (S3b, pure code-motion)
`openai_adapter::generate_stream` was ~1.3k lines doing everything (Joel: "that massive
and improperly named adapter"). Two concerns become modules named for the concern:

- `inference/serving_guard.rs` — the pre-flight that guarantees the single-resident lane
  serves the requested model (waits on the daemon's readiness signal, refuses otherwise)
  and refuses a prompt that alone overflows the served slot (#175). The four predicates
  (`snapshot_guarantees`, `settled_on_another_model`, `unguaranteed_model_refusal`,
  `prompt_alone_overflows_served`) and their test move with it.
- `inference/lane_send.rs` — POST through the lane, ride out a mid-relaunch (connection
  refused / 503 loading) with the bounded retry, classify a rejected status at the seam.
  The header-timeout and relaunch-retry constants move with it.

Behaviour-identical: the blocks are moved verbatim with their measured-history comments,
`self.config.*` becomes the passed config, nothing else changes. Adapter 4191 → 3784 lines.
Remaining carve: `request_body` (JSON assembly) and `sse_parse` (the stream loop).

(cherry picked from commit 1de1961f0db33119a1f0f5b8242ce5f9881efc16)
