+++ core/continuum-core/src/inference/mod.rs
@@ -41,6 +41,7 @@ pub mod handle_store;
 pub mod kv_quant;
 pub mod lane_args;
 pub mod lane_health;
+pub mod lane_send;
 pub mod stream_liveness;
 pub mod lane;
 pub mod lane_pidfile;
@@ -59,6 +60,7 @@ pub mod model_commands;
 pub mod ort_providers;
 pub mod placement_capture;
 pub mod recipe_budget;
+pub mod serving_guard;
 pub mod throughput_expectation;
 pub mod turn_admission;
 pub mod placement_watch;
